Deletador de categorias melhorado, verificando a existencia da categoria

+++ server/src/controllers/deleteCateg.ts
@@ -1,14 +1,16 @@
-import Database from '../db/config'
-import { createResponse } from '../helpers/createResponse'
+import Database from '../db/config';
+import { createResponse } from '../helpers/createResponse';
+import { checkItemExists } from '../validation/checkItemExists';
 
 export async function deleteCateg(id:number, categName:string) {
   try {
-    const db = await Database
-    db.run(`DELETE FROM categories WHERE ownerID=${id} AND name="${categName}"`)
+    const exist = await checkItemExists(categName, 'categories', 'name')
+    if(!exist) throw new Error('Produto não encontrado');
+    const db = await Database;
+    db.run(`DELETE FROM categories WHERE ownerID=${id} AND name="${categName}"`);
 
-    
-    return createResponse(200,'Produto deletado')
-  } catch (err) {
-    return createResponse(404,'Produto não encontrado')
+    return createResponse(200,'Produto deletado');
+  } catch (err:any) {
+    return createResponse(404,err.message);
   }
-}
+};
